[REFACTOR] Repositosy 분리 (#587)
